+++ src/core/controller.ts
@@ -1,6 +1,7 @@
-
+import { PolicyDescriptor } from '../core/policy';
 
 export interface ControllerT {
-    transformPathPart?: { (part: string): string },
-    transformPath?: { (path: string): string }
+    transformPathPart?: { (routeFnName: string, part: string): string },
+    transformPath?: { (routeFnName: string, path: string): string },
+    transformRoutePolicies?: { (routeFnName: string, fullPath: string, policies: PolicyDescriptor[]): PolicyDescriptor[] }
 }